# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/api/health/debug-sentry": {
      "get": {
        "operationId": "HealthController_triggerError",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/api/v1/todos": {
      "get": {
        "operationId": "TodoController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TodoResponseDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Todo"
        ]
      },
      "post": {
        "operationId": "TodoController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/TodoResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Todo"
        ]
      }
    },
    "/api/v1/team": {
      "get": {
        "operationId": "TeamController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ColleagueResponseDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "UserController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PublicUserResponseDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "**AUTH_FORBIDDEN_PERMISSION** — This action requires one of these permissions: {requiredPermissions}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "post": {
        "operationId": "UserController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicUserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "**USER_EMAIL_ALREADY_EXISTS** — Email {email} is already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UserController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicUserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "**USER_NOT_FOUND** — User {id} not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicUserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "**USER_NOT_FOUND** — User {id} not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User soft-deleted"
          },
          "400": {
            "description": "**USER_CANNOT_DEACTIVATE_SELF** — You cannot deactivate your own account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "**AUTH_FORBIDDEN_PERMISSION** — This action requires one of these permissions: {requiredPermissions}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "**USER_NOT_FOUND** — User {id} not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/users/{id}/role": {
      "patch": {
        "operationId": "UserController_updateRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicUserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "**AUTH_CANNOT_CHANGE_OWN_ROLE** — You cannot change your own role<br/>**USER_CANNOT_DEMOTE_LAST_ADMIN** — Cannot change role: this is the last remaining ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "**AUTH_FORBIDDEN_PERMISSION** — This action requires one of these permissions: {requiredPermissions}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "**USER_NOT_FOUND** — User {id} not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthTokensResponseDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "**USER_EMAIL_ALREADY_EXISTS** — Email {email} is already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthTokensResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_INVALID_CREDENTIALS** — Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "**RATE_LIMIT_EXCEEDED** — Too many requests, try again later"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthTokensResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_REFRESH_TOKEN_INVALID** — Refresh token is invalid or has already been used<br/>**AUTH_REFRESH_TOKEN_EXPIRED** — Refresh token has expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthenticatedUserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "**AUTH_TOKEN_MISSING** — No access token was provided<br/>**AUTH_TOKEN_EXPIRED** — Access token has expired<br/>**AUTH_TOKEN_INVALID** — Access token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "Ironclad Initiative API",
    "description": "API documentation for the Ironclad Initiative backend",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "AUTH_INVALID_CREDENTIALS"
          },
          "message": {
            "type": "string",
            "example": "Invalid email or password"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "TodoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "done"
        ]
      },
      "CreateTodoDto": {
        "type": "object",
        "properties": {}
      },
      "ColleagueResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "team": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "role",
          "team"
        ]
      },
      "PublicUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER",
              "MODERATOR"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {}
      },
      "AuthTokensResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {}
      },
      "AuthenticatedUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER",
              "MODERATOR"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      }
    }
  }
}